fingers.today --> 0f
Uses a filesystem overlay and a bind mount to achieve a new view of the
filesystem.

Now use a single templated service file and a new directory structure to
run multiple sites on 0f:

  /data/0f/lib/                         <-- base 0f runtime files
  /data/0f/sites/example.com/lib/       <-- site specific runtime files
  /data/0f/sites/example.com/pages/     <-- site specific pages
  /data/0f/sites/example.com/workdir/   <-- workdir for FS overlay

Sites are able to customise their instances by creating their own
versions of files like CSS and view templates. These files are overlayed
on top of the base 0f files and mounted read-only at:

  /run/0f/example.com/

Then the pages directory is read-write bind mounted to:

  /run/0f/example.com/pages/