Created and last modified times using xattrs
I decided to set the created time to the modified time if it doesn't
exist (which would mean on first web view). This allows for creating
pages on the filesystem using different means than only through 0f.

In the same way, last modified time is read from mtime, if it is not
already set.