Add dynamic keyboard shortcuts for links
When authenticated, new keyboard shortcuts are available. When pressing
an alphanumeric key, navigate to the first encountered link in the
document that starts with that character.

To avoid conflicts, other shortcuts have changed:

  / naviagtes to /
  . naviagtes to /private/
  a no longer brings up the add page modal (right-alt-a still does)

These shortcuts are not enabled if not authenticated. Too confusing if
you don't know what's going on and seems unlikely anyone else would use
them anyway.